:sparkles: Migrated some of the models
I migrated some of the models already - made the actual models schema-wise (and of course DB wise), and migrated some of the data over. Also updated some of the DB TS types to tsome of the new schemas (mainly just the ones that i already migrated the data for, though). Migrated tokens and networks and strategies and users. Next gonna do protocols which should be fairly easy. tThen the harder bits (addresses, functions , args, flows, etc)