:lipstick: Improved responsiveness of styling for buttons
Improved the CSS of the buttons and some other stuff to adapt for smaller screens. Still want to make the custom hook, make the text dynamic for create vault button, make the same button fix for the dropdown items (?), and also think of how i can make the catagoryu links on the header more compact.