:sparkles: New user auto signup now works.
The new user auto signup on new address connection now works. Although i have noticed a slight bug in the username rendering - seems like it does not render on the first connection... not urgent but investigate when you have time. It seems to update on concurrent address changes, and also the PFP does seem to change - easiest way to debug would be to look at the difference of how the PFP and how the username states are hcanged, and also how thyere handled?