:bug: New user signup now deff executing only once.
I made a commit yesterday saying i fixed it but was wrong - It wasnt spam-signing me up, but it still signed me up twice each time, and also didnt update the users context so it would re sign me up on each rerender... I fixed two issues now; one was the fact that it wouldnt refresh the context as i said, so i just had it refresh it within the async signup function. The second issue was that, the useEffect was (obv) created per consumer of the custom hook. So if there was 3 consumers of the hook, it would run 3 times in parallel... i had to insert an internal zustand store in order to persist a ``isSigningUp`` state across all consumers, which seems to now work!