:sparkles: Added abillity to change username
You can now change the username. Note that there is still issues with the username update, for some reason. When i change the username, it does not hcange on the interface... IT seems to update in the cntext itself, but no rerender (or even comaprsison) of users happen within the custom component. IF i  close the modal and reopen it, it works.