:sparkles: Added Strategies Slideshow
Took a while. Didnt make my own generic component because it wouldve been too difficult (i tried and it was really laggy and ugly)... instead i battled with react slick to append the dots in a way that shows the active one and also the arrows (which  are still not aligned proprely to my taste). But it looks OK for now, i did notice it is potentially slowing the app down so ill try building and see if it does, and try to opt it