:sparkles: Modal routing, started strat modal work
The modal routing now works w a catch - the route itself isnt handling pushing, but some other parent. I couldnt get it t o work unforutantly as it seems that when the dynamic route is in a layout group (Which is required for this to work), the component simply refuses to render with no hints as to why... I starte dworkig on the strategy modal design - added title, intiial flexing of the boxes, and the profile section. Tomorrow ill work on adding actual functionality to them, adding web3 integration (which is libs work as well, should be pretty fun), and once i finish and it looks good ill start working on the migration of the steps and the entire infra for it then lay it out at the bottom. Might fuck around and put some placeholder chart on the top box in the meantime until i make that service