:bug: Fixed all hydration bugs, hot reload works
Holy fuck this feels so much better now. Not only did i fix the hot reload errors but also the annoying hydration errors i was getting all the time when the page loads... The issue with the hydration was indeed the mount. Though i already tried to fix it on the ```useYCUser```, it kept going. So what i did was (got the idea and then saw it on some article and was like yea that makes sense), i simply wrapped the entire app in a mounting requirement. (The entire client body that is - inside the ```WagmiConfig``` that wraps everything.), now all the stuff that may require different renders than the server (i.e all the client shit im doing, which i do not want to repeat a mounting requirement in every single one) are requiring a mount to render, so theres no mismatch on the initial paint. The hot reload issue i think was caused by this + some mismatch in how i structred the layout html... had to put everything within the body rather than outside. The z-index of the ``ModalProvider`` feels a bit cluncky now (it didnt work when i gave it unlimited z-index, but only did when i ordered it to the bottom of the siblings in the root layout), but ig its okay. For cleaness, i might want to shove the mounting requirmeent in its own component - like ```<Client></Client>``` and wrap the entire app in there. It does not realy matter, just clean and makes it more explicit that i want all of its children to render only on the client (and hence require the wrapping mount directly in there).