:bug: Fixed hot reload again. Afraid to restart terminal
mad this is spooky. i fixed it last commit then restarted and it fucking came back again the piece of shit. Now i excluded node modules from the babel loader in my next config and it hot reloads now, but i am so scared of restarting the terminal session lol