:construction: Half way done through chart, InfoProvider
im mid way w the chart - it works, shows the data nad looks OK. I want to add a custom tooltip (the infoprovider) and timeframes (which should be hard as this is just filtering the data bsaed on time). AS for the infoprovider, it sort of works. It works when i input the position manually ot be above the child - but my function to calc based on direction wasnt working. Wil lfix it tomorrow and also fix another bug where the wrapped text i put in does not work i think its either not taking the fontsize randomly for whatever reason or perhaps its not detected as string so its returning it plainly (Try returnign dummy shit there to see if it is).