:bug: Fixed toJSON of libs, started working on FE step
Fixed (atleast for now) the issues with the toJSON / stringification of the libs. I have a generic method which iterates over the fields and parses them manually depending on the current depth (that can be manually specified). If the depth is lower than the specified depth (i mean deeper) and the field is an object which has an ID field, then it will return it. The good thing is i am pretty sure i can override this function in the individual classes as well, so if a certain class needs some unique parsing method for its toJSON then i can implement it directly on it