:sparkles: Draggable canvas, drag/wheel/zoom, bounds
Made the draggable canvas - U can drag, wheel it, and zoom, with custom zoom constraints and position constraints to match parent. One thing that is kinda meh for now is the zooming setter but ill come up w a better looking solution once i actually need that, but its not that bad. Once thing i need to do is set the height/width proprely also of the canvas - min-h-full for some reason acts as a max height, and trying to do h-calc... doe snot work. Actually for now ill just add padding since it will be a flex anyway, but yea might wanna do it...