:sparkles: Canvas sizing auto-calc, nodes layout on canvas
Made the ``graph()`` function return the canvas' desired dimensions and entered that to the canvas itself (with min w and h still set to full). The nodes seem to layout nicely now with no lags and centered somewhat correctly. Still have an issue with the bounds being incorrect after zooming not in the center