:sparkles: Re-migrated steps with flows
Remigratred steps in DB to ones with flows. I decide to save them as JSONs in steps's flows array and not as references to IDs. Gonna migrate it in the Steps libs now and a bit later on in the entire data structure - no need to save an ID for a token bundled  with a direction. Infact i can just save them in 2 seperate columns in functions for instance (inflows outflows) and just reference a token ID there