:sparkles: Persistance Strategy Draft Stire
Made the zustand strategy draft store with persistance. Only issue is that it does not write on initiallization which is weird... But anyway not too big of an issue as it works if i shove it in a useEffect. Worst case i make a custom hook for it if i really want but really it will only be used in 1 place so no biggie