:bug: Changes now cause graph rerenders
Before it didnt work. The issue was that i misused the useEffect along with the shouldGraph() function... it returns a boolean, yet i was putting it in the dependecy effect as if returning a boolean counted as a state change. If it were to return true, then there was some change and it would return true again - It would not count as a change. I therefore now just JSON stringified it in the dependency array. Might make a shorthand function for it or even a custom hook (useStepRegraph()). And maybe i will use the shouldGraph function in the future to avoid rerendering based on ANY change, but only important ones