:bug: Steps initiallization now works
It was a bug within the custom ``useSteps`` hook. The useEffect which is supposed to initiallize the rootstep of a strategy was buggy when it was already initiated beforehand. I fixed it by checking whether it exists already and only doing the new instance if it wasnt. Because it was creating a new instance on an already graphed root step, so thenew instance's positions were 0,0 all the time.