:bug: Fixed a bug with custom sized nodes when manually resizing them
THe flow should have been like: User Resizes => Default DImensions get assigned => component renders => useEffect runs to resize to it's dimensions => components resizes again to the new dimensions. YEt the issue was that, once a user manually resized  the component, it was considered manuallyResized, and the resizing on the useEffect in the Base Node did not do anything. I fixed it by adding a true to the parameter that says if it is a mnaual resize or not. This should not hurt the dirty/clean resizing flow as this resize only resizes the dimensions themselves to the existing size proprety of the step, so that there is no clashes.