:bug: Fixed new tooltip behavior only working after rerender
I was using this custom hook i made for getting the element of an ID, yet it was acting inconsistently - Some specific rerender was required in order for it to get the actual element. This was because (it seems like it, atleast) i was memoing it and having it change every time the ID changes - Whilst the logic is theoretically correct, this was often fired before the DOM fully mutated. And therefore, even if the ID was correct, the DOM element was not yet present (atleast with it's ID), so it was returning a nullish value. I fixed it by, instead of having a single memo in the custom hook - having a state initiallized to undefined, then running a ``useLayoutEffect()`` every time the id changes. ``useLayoutEffect()`` runs synchronasely after the DOM has fully mutated, which makes sure the element we are trying to access has had it's ID set