:sparkles: Harvest
This was really easy lol i didnt even make a config component. I realized that now that i do not assume u can do everything form anywhere like in the last fronend, you will only ever have a single harvestable position (or maybe a few that will relate to a specific function). Atleast thats what i think for now, if i reach a point where it seems like there can be multiple, ill make a config component. But for now it just finds one that depends on the parent and auto chooses it