:bug: Fixed Dropdown Menu Ref Not Working
The issue was with the infoprovider. The issue was now that there must be a ref provided by the infoprovider in order for it to work... when i changed it to return the buttonRef, it only worked when there was NO wrapping info provider. This is becuase it uses a callback ref, which handles ref setting/cleanup itself. So if i was wrapping it in an info provider, it would assign the ref elsewhere, and the ref is not accessible through props (what i thought would prevent this in the past - because it just accesses the nearest parent's props, and not the underlying div's). So what i did was add a refSetter prop to infoprovider which would set the ref somewhere, and added a useEffect to the dropdwn to set it's own prop of refSetter to set to the dropdownBtnRef, and its now working