:sparkles: Full harvest, unlocked funcs logic from seed => tree.  medium,  small harvest configs
Made the full logic to allow the seed to unlock the tree's functions. I think in the near future i may also need to add some more logic to handle removal of children if user went back to edit seed steps and changed behavior of it. But looks good, can handle removal of children and etc etc